Move observability package to system package
