+++ backend/internal/system/observability/adapter/console_adapter.go
