+++ backend/internal/system/observability/adapter/console_adapter_test.go
