+++ backend/internal/system/observability/interface.go
@@ -19,10 +19,10 @@
 package observability
 
 import (
-	"github.com/asgardeo/thunder/internal/observability/event"
-	"github.com/asgardeo/thunder/internal/observability/publisher"
-	"github.com/asgardeo/thunder/internal/observability/subscriber"
 	"github.com/asgardeo/thunder/internal/system/config"
+	"github.com/asgardeo/thunder/internal/system/observability/event"
+	"github.com/asgardeo/thunder/internal/system/observability/publisher"
+	"github.com/asgardeo/thunder/internal/system/observability/subscriber"
 )
 
 // ObservabilityServiceInterface defines the contract for the observability service.
