+++ backend/internal/system/observability/service_test.go
@@ -22,8 +22,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/asgardeo/thunder/internal/observability/event"
 	"github.com/asgardeo/thunder/internal/system/config"
+	"github.com/asgardeo/thunder/internal/system/observability/event"
 )
 
 // setupTestService creates a test service with controlled configuration.
