+++ backend/internal/system/observability/subscriber/registry.go
