+++ backend/internal/system/observability/subscriber/registry_test.go
@@ -21,7 +21,7 @@ package subscriber
 import (
 	"testing"
 
-	"github.com/asgardeo/thunder/internal/observability/event"
+	"github.com/asgardeo/thunder/internal/system/observability/event"
 )
 
 // mockSubscriber is a mock implementation of SubscriberInterface for testing
