+++ backend/internal/system/observability/subscriber/subscriber.go
@@ -20,7 +20,7 @@
 package subscriber
 
 import (
-	"github.com/asgardeo/thunder/internal/observability/event"
+	"github.com/asgardeo/thunder/internal/system/observability/event"
 )
 
 // SubscriberInterface is the interface that all event subscribers must implement.
